Load the whole config file all the time
This avoids iterating through all config items of each modules. On my
system, this cuts the vlcrc load time in half, and is also far simpler.
The loader is essentially ignoring the config section, which reflects
the architectural "limitation" of the VLC configuration.

This also avoids loosing settings during VLC upgrades if an option moves
from one module to another one (which did happen in the past).