skins2(Win32 and Linux): Don't refresh a window forcefully
A good practice is to invalidate the window and leave it to the OS to decide if a repaint is or not needed.

As a side effect, this fixes some alternative task switchers (alt-tab) on WinNT (like TaskSwitcher or ATTv) that displayed a black rectangle instead of the preview of the skin, because the skin engine was not doing things the usual way.