relative_URI memory leak cleanup
If the the link is relative, psz_uri is dynamically allocated a string of the
absolute URI (which is also returned by the function).   However, if it isn't,
NULL is returned and nothing is allocated.   segment_New performs a strdup, so
the pointer is not saved.   Thus, check if the allocation occurred and clean up
if needed.

Signed-off-by: Jean-Paul Saman <jean-paul.saman@m2x.nl>