rtp sout: send in-band vorbis/theora configuration packets
This is an alternative to SDP to convey the codec configuration
necessary to decode the stream. The configuration is sent only once at
the beginning of the stream; at least you won't be tempted to waste
bandwidth with retransmissions. This is a ad-hoc hack, but it's not too
bad or pervasive.