access/v4l2.c: Tell vlc what framerate the device delivers
The framerate as specified by the user or detected by the v4l2 access input was
overwritten by the rawvideo decoder. The module did not specify:

	es_fmt.video.i_frame_rate
	es_fmt.video.i_frame_rate_base

Their default values are 0. The rawvideo decoder detects that there is no framerate
specified and resorts to its own default (fps 25.0). Through setting framerate and
framerate_base in the v4l2 es format the correct fps is communicated to vlc core.