Broadcom CrystalHD decoder support
This new module works with BCM70015 Crystal HD video decoders
It handles hardware decoding of H264, VC-1 and mpeg2

So far, it only works on GNU/Linux with the appropriate driver and
library. Windows version is still limited.

This is the first part, since mpeg4, interlaced and BCM70012 are not
working yet.