stream_filter/httplive.c: b_cache determined in wrong place
The logic to determine if caching of segments is allowed or denied was flawed. It
only determined it onces after the playlists were loaded. In case of a meta playlist
the logic would only indicated the caching value for the last added HLS stream.

The fix is to determine p_sys->b_caching (caching segments allowed or not) when switching
segments or HLS streams (adaptive bandwidth).