Qt4: remove custom X11 display with --x11-display
In older VLC versions, we had to use the same X11 display for the
window provider (the UI) and the video outputs to avoid Xlib aborts.
We do not need this anymore. Morever, the XCB outputs can detect the
error and fail safe.

To change the UI display, the DISPLAY environment variable is used.
The only real use case for --x11-display is output to a different
display than that of the UI. This change restores this use case,
and simplifies the Qt4 interface.