stream_filter/httplive.c: Do not crash when Prefetch() fails
If Prefetch() decides that playback cannot commence it returns VLC_EGENERIC. The
code path in Open() tries to cleanup by calling Close(). However at that point the
hls_Thread() is not started yet and this causes a crash in Close().

The resolution is to not call Close() but cleanup the parts that have been created
in the Open() function itself.