Win32: revert to old rc interface
See #4455. The lua interface merely polls, reads and writes on file
descriptors 0 and 1, and expects it to work. There is no support for
spawning a console and redirecting/getting file descriptors to it, and
file descriptors are mixed with socket descriptors and polled, which
won't work on windows. Until something is done about it, revert to the
old rc interface