Rationalize video output priorities
This tries to prioritize stuff a little better and more uniformly:

 - less than  20: console (text)
 - less than  50: plain frame buffers
 - less than 100: no embedded windowing
 - less than 150: SW scaling
 - less than 200: SW chroma conversion
 - rest: good stuff :-)

This removes multiple plugins with identical priorities. On Windows,
native outputs are now priorized over XCB (as there now is a Win32 port
of XCB). The ordering amongst native Windows outputs is preserved.
On X11, GLX is put ahead of plain as it supports scaling. SDL and
DirectFB are demoted as they cannot be embedded.

Old order:

 - XCB XVideo
 - Windows Direct3D (Vista+)
 - Windows DirectX
 - XCB X11
 - Windows Direct3D (XP-)
 - Windows Direct2D
 - DirectFB, SDL
 - XCB GLX
 - Linux FB
 - Windows GL - Windows GAPI
 - Colored ASCII Art
 - ASCII Art - Windows GDI

New order:

 - Windows Direct3D (Vista+)
 - Windows DirectX
 - Windows Direct3D (XP-)
 - XCB XVideo

 - Windows Direct2D
 - Windows GL
 - XCB GLX

 - Windows GAPI
 - Windows GDI
 - XCB X11

 - SDL
 - DirectFB
 - Linux FB

 - Colored ASCII Art
 - ASCII Art