Disable bandlimited resampler for the time being
It is known to crash due to invalid buffer handling. On Linux, I get
memory accesses before the start of the audio buffer. It also crashes
on Windows. Until someone cares to fix it, and I think it really never
worked correctly, lets use the "ugly" resampler, which is not too ugly.

This should fix #3631, #4366, #4491 and possibly others.