DTV: work-in-progress Linux DVBv5 input
This is incomplete and hence not automatically compiled. Conditional
Access, DVB-S and scanning are missing. DVB-T and ATSC are not tested.
DVB-C works.

It is not clear to me what the problem is, but it seems VLC does not
like the Linux DVB kernel demux. Only budget mode seems to work
correctly at the moment.