sout_display: remove always NULL vlc_object_find()
The stream output is not a decendent of the input in VLC 1.2. So this
plugin does currently not work at all. This commit only removes
non-working code; it does not break or fix anything as such.

In previous versions, this code could cause crashes when #gather was
used, as the input object would change behind the back.