stream_MemoryNew: remove vlc_object_find()
As for the previous commit, this seems more harmful than helpful.

Due to the UTF-16 subtitle hack present in the stream layer, this
commit will break in-memory UTF-16 subtitles character encoding if
UTF-8 autodetection was disabled and unless UTF-8 is the selected
encoding. On the other hand, this commit fixes subtitles encoding
if there is more than one subtitles stream in the whole input.