Freetype: do not use fontconfig on Win32
Fontconfig is great, but we have too many issue, notably the fact that
the fontcache isn't save correctly in some locations, notably on the
network. It also rebuilds the entire cache, each time you add ONE font.

Therefore, this commits uses Win32 EnumFontFamilies, and then gets the
font path with the Registry + concatenation of CLSID_FONTS and then back
to freetype.

However, this commit is broken in at least 3 ways:
 - it does not use the index to select the font,
 - it will break on localized fonts settings (Eastern),
 - I don't check the Script yet,

 I will try and fix in the future.

Ref and close: #4470, #4401, #4137, #3950, #3208