Require an aout instance to create an aout input
The input resource creates the aout instance with the input manager as
its parent object.

If this fail, the aout input should not be created at all. This removes
the fallback to an aout instance with the decoder object as parent.
This would potentially crash as the decoder is shorter-lived than the
aout instance.