V4L2: Allow device to be specified with option.
Currently the V4L2 device to be used can only be set via the media URL.
This patch allows the :v4l2-dev option to set the device as well.
Setting the device via the media URL still takes precedence.

This is useful if a program auto-detects the V4L2 device so that the
media URL only needs to contain e.g. the tuner settings.

Signed-off-by: Rémi Denis-Courmont <remi@remlab.net>