Mark other symbols as hidden
Unfortunately, libtool is yet incapable of doing marking symbols as
hidden (not exported) at compilation time.  So lets do it manually.
According to the GCC documentation, this allow some compile-time
optimizations depending on the platform.
This was already done by Damien earlier (and removed during the
plugin headers cleanup).

The check is disabled on Windows, as the compiler accepts the flag,
but generates lots of warnings during compilation.