Remove endianess and type sizes from plugin name
On the one hand, plugins from different architectures cannot be mixed
in the same installation directory as they have the same names.
On the other hand, endianess and type sizes is way insufficient to
discriminate architectures (e.g. armel and i386 look the same).

So this was totally useless. And it did not need to be formatted at
run-time either.