Add support network URI schemes for KDE
Warning: If credentials are required, we are going to have to assume
that KDE passes them in the URL. I have not tested if this was the case
or not. This is potentially nasty for those protocols whereby VLC does
not currently implement any login dialog. But ultimately, I would
consider this a KDE bug (if applicable).