Removed useless vlc_thread_set_priority() calls in macosx code.
The thread are already created with the right priority. If not, then
vlc_clone() should be fixed instead.
Also, the vlc_thread_set_priority() calls on the main interface thread
seem bogus, as the target thread is dedicated to monitor the libvlc
death and is not the main running thread.