Added YUV(planar)->RGB conversion support to opengl using a fragment program.
All planar YUV are supported (only YV12, I420, I422, I444 were tested).

For now, the YUV is supposed to follow BT-709 when video height is higher
than 576, and BT-601 otherwise. The YUV range is supposed to be TV, and
the RGB one is supposed to be full range.

In the current state, the behaviour is unchanged on macosx (ie YUY2
is used instead), and when opengl ES is requested (RGB 16 bits).