Turn --services-discovery into a string advanced option
Modules list was intended for plain static module names. It does not
work right for most SD plugins, including:
 - LUA (dependent on string chain),
 - udev (submodules),
 - media directories (submodules).

It works but is totally useless for those plugins that are fast,
including XCB screen capture and Windows optical drives.

For network-based discovery and MTP, it could be used to prefetch
items. But this is not necessarily a good idea as it sends extra
network traffic and consumes memory even when the services discovery
is not used. Furthermore, podcast can prevent Qt4 and Skins2 UI from
starting, as it depends on HTTP which depends on libproxy which
depends on Xlib.