Deprecate playlist_IsServicesDiscoveryLoaded()
With the introduction, initially for LUA SD, of the chain syntax to
the services discovery, there is no way to canonicalize the name of
a services discovery object. Thus, they cannot be compared reliably,
and therefore, playlist_IsServicesDiscoveryLoaded() does not operate
correctly. I won't remove it today because too many interfaces call it.
But you should know that this function does _not_ actually work, and I
see no way to fix it (other than remove LUA SD).

Futhermore, this function is fundamentally prone to a race condition,
when it is used to toggle an SD (this is admittedly a minor problem).