omxil: Only set the buffer size if it is smaller than what the codec suggested
This is required on HTC Sensation. That device uses a special
tiled pixel format, which probably contains a bit of padding, too.
The size that the omx codec suggests for the output port is bigger
than what VLC calculates, and we thus should keep the larger value
instead of overwriting it with what we calculated ourselves.

Signed-off-by: Rafaël Carré <funman@videolan.org>