aout: clear out input and output FIFO simultaneously
This occurs when the output is really late (typically as a side effect
of glitches in pause/resume). Both FIFOs  are now both protected with
the same lock, so there is no room for race conditon between the two
identical checks anymore. So we can handle both checks in one go.