aout: do not check for PTS discontinuity in packetizer
Continuity of the two FIFOs is maintained, as is consistency with the
input date. PTS discontinuity only seems to occur due to rounding
errors in the filters. It should be better not to drop samples, but
simply fix the buffer PTS and length.

Ideally, resamplers would use date_Increment() so they do not introduce
rounding errors. But first the block disconuity flag should be checked
for correctness during filtering, so that resamplers can reset the date
in case of hole.