Remove unused/unusable decoder_DeleteAudioBuffer()
Audio buffers are destroyed downstream of the decoder, so this
function could never be used. In practice, the audio filters, the
audio output (or the encoder) already use block_Release() to delete
buffers allocated with decoder_NewAudioBuffer().

If "direct rendering" is ever implemented in audio output,
block_Release() custom buffer deletion can be used.