Use hash instead of arobas for seekpoint (fix #5022)
Instead of ignoring the URI anchor (if present), the input now parses
it for start and stop seek points. This is non-standard - anchors are
normally names found in HTML document. But at least, it does not
interfere with parsing real-world URLs (see the bug #5022 for example).

Also contrary to stream output chain, the hash symbol should be free of
shell quoting issues when used this.