omxil: Don't trust a qcom decoder claiming to output 420 planar
Also move the color format definition to the header, making
it available to omxil.c.

In practice on Nexus One with stock firmware, it is NV21
even though the decoder says it's planar. (On CyanogenMod with
OMX decoder built from source, the decoder returns
OMX_QCOM_COLOR_FormatYVU420SemiPlanar as pixel format, though.)

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>