win64: read/write x64 registry, not the x86 one
InstallDirRegKey is not affected so we have to emulate it

plugins are now registered correctly, and both 32-bits and 64-bits
versions can be installed side-by-side without the installers screaming
However there is an obvious conflict in the menu, desktop icon, and
file / actions / context menu associations.

In my testing the 32-bits version was selected as default, and it was
also the last one installed.