Win32: do not set _fmode
Setting the translation mode is the responsibility of the program. A
library must not interfere with that. It could perhaps be added to VLC.
However it should not be required as we now force binary mode by
default in vlc_fopen() and always specify an explicit mode when calling
fopen() directly. Also it would make VLC behave differently than other
LibVLC applications, which may raise really weird bugs.