Fix thread library check
Don't use ac_cv_search cached value because it could contain
"no" or "none required" which aren't valid link flags

AC_CHECK_LIB(main, xxx) did not exactly search for main symbol, it
searched for the presence of the library.
Instead we look for a required symbol.

We don't use e.g. pthread_mutex_lock because some pthread symbols
are available in glibc, but not all.

Not finding pthread library is not an error, because at least OS/2 and
symbian don't need it. (We skip the check on Windows already)