omxil: Build the iomx wrapper code as a separate shared library
The separate shared library needs to be loaded into the process
before using the iomx module, preferrably by the java part
of the android application.

This allows building the iomx glue code for a number of different
ABI versions, and additionally avoids linking the main VLC
library to any private symbols that might make it unloadable.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>