omxil: Handle sentinel buffers in DecodeAudio, too
These buffers were introduced to avoid a potential deadlock
when the codec has issued a PortSettingsChanged event but
not returned any buffers to the client, making the DecodeVideo
function hang waiting for an input buffer to use, before
handling the reconfigure. The same handling needs to be done
for DecodeAudio too, otherwise it will try to use the
sentinel buffer as a normal one.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>