auhal: always set audio device callback
Previously, the audio device callback was only set when you call probe.
But in the case you reload the module (for example when switching from
encoded to unencoded output over the UI menu) probe was not called again.
This resulted in a crash when we want to delete the callback while closing.

Signed-off-by: Felix Paul Kühne <fkuehne@videolan.org>