omxil: Add a prefix to the iomx function symbols
On some devices (apparently froyo and earlier), some OMX core
(with sw-only codecs) already is loaded into the process, and
dlsym(RTLD_DEFAULT) can just as well return functions from that
one instead of the ones from the iomx wrapper that we've loaded.

This makes sure we really get the functions we want.

Signed-off-by: Rafaël Carré <funman@videolan.org>