omxil: Set the module priority to 0 for iomx
The iomx module is only enabled when the user has requested it
(via settings). Currently, when the iomx module has priority 80,
it is always preferred over avcodec.

For other platforms, where one would only enable omxil if one
explicitly wants to use it, this is ok, but for android, we
ship the iomx module but allow users to enable it at runtime
if they want to.

If it isn't enabled, the priorities are currently overridden by
adding "--codec avcodec,all" to the command line, but this makes
avcodec have precedence over other decoder modules that otherwise
would have higher priority (e.g. a52).

By setting the priority to 0 for the iomx version of this
module, we can more easily enable it only when asked for,
by adding "--codec iomx,all", without affecting the relative
priorities of all other codec modules.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>