Use newer theora API and fix crash
Close #3841

This updates the theora module to use the newer API introduced in Theora version 1.0.
It solves a crash on windows that was caused by reading past the end of the YUV structure due to an incorrect line count.
This has been fixed by using the minimum line count common to both the source and destination of the pixel data.
The new ycbcr structure in the new API makes the picture copy much more straightforward.

Modified-by: Jean-Baptiste Kempf <jb@videolan.org>
Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>