dbus: Prevent dictionary from growing out of control
For each item that was appended or deleted, the ("Tracks", NULL) pair
was inserted (duplicated over and over) in the tracklist_properties
dictionary. The dictionary was inserting it in the same position and
it assumed it had hash collisions, triggering an expansion at every
insertion of that pair.

Check if the key is present before inserting in the dict again, at
playlist_item_append, playlist_item_deleted events.

Signed-off-by: Ilkka Ollakka <ileoo@videolan.org>