ALSA: add configuration item for channels, remove broken autodetection
There is no reliable way to determine the channel configuration of a
certain ALSA device. The ALSA operates at too low level for that.
So add a configuration option with default to stereo (as in previous
VLC versions). However, the user will not need to re-enable 5.1 every
time.

This commit also removes all the fancy device choices. The static
choices are overriden dynamically by the UI anyway. So the only use of
the list was in vlc --help.