Enable some FPU optimizations via standard pragma by default
If GCC -ffast-math is not supported, some floating point optimizations
are still possible through more standard means:

"FP_CONTRACT on" is a subset of -funsafe-math-optimizations.

"FENV_ACCESS off" corresponds to -frounding-math.

"CX_LIMITED_RANGE on" corresponds to -fcx-limited-range, but is not
included as VLC does not use complex numbers anywhere yet.