Linux DVB: run-time fallback from 5.5 to 5.0 frontend probing
This enables running VLC to run on a kernel that does not support
version 5.5 of the Linux DVB API (i.e. Linux < 3.3), even if VLC was
compiled with more recent kernel headers.

Unfortunately, this is of limited use. In practice, the running kernel
version is more likely to be newer than the kernel headers, than to be
older. In that case, VLC will print an explicit error.