decoder: pause aout at creation in corner case (fix #6369)
If pausing the decoder races with setting up the audio output pipeline,
the decoder can end up already paused when the audio output is created.

The audio output insists on a strict start -> (pause -> resume ->)*
(pause ->)? stop series of state transition. So it must be paused
explicitly in this corner case.

I have not investigated whether this problem affects the video output
but this seems very likely too.