skins2: resize new layout first when switching between layouts.
When switching layouts the windows was resized to the default state at design time.
Now the layout being switched to is first resized to the size of the current layout.
Switching layouts now nolonger results in the layout being restored to its default size,
instead it keeps the size the user inteded it to have.