Qt: use a QStyle to draw the seek slider
This should be much cleaner than drawing ourselves inside the paintEvent
of the widget. Furthermore it is easier to fallback to the native slider
available in pre 2.0.
The use of a QStyle also fixes the clipping issue where the handle was
cropped while being at the start or the end of the slider.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>