PulseAudio: adjust overall latency, not (only) the buffer size
Without this flag, PulseAudio will tweak the buffer parameters, but
will not account for hardware latencies. As a consequence, the live
caching parameter is not really followed - the latency is longer.

Longer latency becomes a big problem when trying to synch with another
live source (such as a V4L2 video capture device). Then audio is too
late and gets discarded or messed up at output.