Undo misuse of module_config_t.b_dirty
This fixes a bug whereby the configuration would not be saved because
Qt4 cleared the b_dirty flag instead of minding its own business.

The core determines whether the value of a configuration item as changed
using b_dirty. b_dirty is _not_ meant to signal changes in choices list.
Configuration callbacks blindly set b_dirty, so the UI can assume that
choices changed. Besides those callbacks should be removed as they do
not follow the locking model of the configuration and can cause crashes.