omxil: Don't require SetAudioParameters on the decoder output port to succeed
It's mostly enough to call it on the input port, while we later get
back what the decoder itself set on the output port (based on the
input port and codec config data), so if this fails, just ignore it.

This is required for using google SW audio codecs via OMX, which is
useful mostly for testing. (Using them still requires a few more
hacks though.)

Signed-off-by: Martin Storsjö <martin@martin.st>