omxil: Allocate the sentinel buffer dynamically
This avoids issues if there's more than one sentinel added
to a FIFO at the same time, since the FIFO is built as a
linked list.

Also use a custom buffer flag to mark sentinel buffers, to
avoid issues if a codec would set the EOS flag on returned
input buffers.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>